Creating seperate classes for objdump and llvmobjdump parser
